Divied file by cateroy class
