+++ VendingMachine/VendingMachine/Soda.swift
@@ -1,0 +1,18 @@
+//
+//  Soda.swift
+//  VendingMachine
+//
+//  Created by oingbong on 2018. 9. 20..
+//  Copyright © 2018년 JK. All rights reserved.
+//
+
+import Foundation
+
+class Soda: Beverage {
+    private var sodium: Int // 나트륨함량 (mg)
+    
+    init(sodium:Int, brand:String, capacity:Int, price:Int, name:String, dateOfManufacture:Date, manufacturer:String) {
+        self.sodium = sodium
+        super.init(brand: brand, capacity: capacity, price: price, name: name, dateOfManufacture: dateOfManufacture, manufacturer: manufacturer)
+    }
+}
